#165 Move SaveGame properties to SaveGame
 <organisation
