+++ src/main/java/nl/tudelft/ti2206/group9/ShaftEscape.java
@@ -70,7 +70,7 @@ public class ShaftEscape extends Application {
 
     /** Creates the savefile directory. */
     private static void createSaveDirectory() {
-        final File saveDir = new File(State.getDefaultSaveDir());
+        final File saveDir = new File(SaveGame.getDefaultSaveDir());
 
         // if the directory does not exist, create it
         if (!saveDir.exists()) {
